Put some helpers in a separate composable
